+++ cratedb_toolkit/adapter/rockset/server/dependencies.py
@@ -9,5 +9,5 @@ from cratedb_toolkit.util.database import DatabaseAdapter
 @lru_cache
 def database_adapter() -> DatabaseAdapter:
     # TODO: return config.Settings()
-    cratedb_sqlalchemy_url = os.environ["CRATEDB_SQLALCHEMY_URL"]
-    return DatabaseAdapter(dburi=cratedb_sqlalchemy_url)
+    sqlalchemy_url = os.environ["CRATEDB_SQLALCHEMY_URL"]
+    return DatabaseAdapter(dburi=sqlalchemy_url)