+++ cratedb_toolkit/cfr/cli.py
@@ -29,10 +29,10 @@ def sys_export(ctx: click.Context, target: str):
     """
     Export CrateDB system tables.
     """
-    cratedb_sqlalchemy_url = ctx.meta["cratedb_sqlalchemy_url"]
+    sqlalchemy_url = ctx.meta["sqlalchemy_url"]
     try:
         target_path = path_from_url(target)
-        stc = SystemTableExporter(dburi=cratedb_sqlalchemy_url, target=target_path)
+        stc = SystemTableExporter(dburi=sqlalchemy_url, target=target_path)
 
         archive = None
         if target_path.name.endswith(".tgz") or target_path.name.endswith(".tar.gz"):
@@ -58,9 +58,9 @@ def sys_import(ctx: click.Context, source: str):
     """
     Import CrateDB system tables.
     """
-    cratedb_sqlalchemy_url = ctx.meta["cratedb_sqlalchemy_url"]
+    sqlalchemy_url = ctx.meta["sqlalchemy_url"]
     try:
-        stc = SystemTableImporter(dburi=cratedb_sqlalchemy_url, source=path_from_url(source))
+        stc = SystemTableImporter(dburi=sqlalchemy_url, source=path_from_url(source))
         stc.load()
     except Exception as ex:
         error_logger(ctx)(ex)
@@ -121,7 +121,7 @@ def job_statistics_collect(ctx: click.Context, once: bool, reportdb: t.Optional[
     """
     import cratedb_toolkit.cfr.jobstats
 
-    address = DatabaseAddress.from_string(ctx.meta["cratedb_http_url"] or ctx.meta["cratedb_sqlalchemy_url"])
+    address = DatabaseAddress.from_string(ctx.meta["http_url"] or ctx.meta["sqlalchemy_url"])
     report_address = None
 
     if reportdb:
@@ -164,7 +164,7 @@ def job_statistics_view(ctx: click.Context, reportdb: t.Optional[str], deanonymi
     """
     import cratedb_toolkit.cfr.jobstats
 
-    address = DatabaseAddress.from_string(ctx.meta["cratedb_http_url"] or ctx.meta["cratedb_sqlalchemy_url"])
+    address = DatabaseAddress.from_string(ctx.meta["http_url"] or ctx.meta["sqlalchemy_url"])
     report_address = None
 
     if reportdb:
@@ -195,7 +195,7 @@ def job_statistics_report(ctx: click.Context):
     """
     import cratedb_toolkit.cfr.marimo
 
-    address = DatabaseAddress.from_string(ctx.meta["cratedb_http_url"] or ctx.meta["cratedb_sqlalchemy_url"])
+    address = DatabaseAddress.from_string(ctx.meta["http_url"] or ctx.meta["sqlalchemy_url"])
     os.environ["CRATEDB_SQLALCHEMY_URL"] = address.dburi
     cratedb_toolkit.cfr.marimo.app.run()
 
@@ -212,7 +212,7 @@ def job_statistics_ui(ctx: click.Context):
 
     import cratedb_toolkit.cfr.marimo
 
-    address = DatabaseAddress.from_string(ctx.meta["cratedb_http_url"] or ctx.meta["cratedb_sqlalchemy_url"])
+    address = DatabaseAddress.from_string(ctx.meta["http_url"] or ctx.meta["sqlalchemy_url"])
     os.environ["CRATEDB_SQLALCHEMY_URL"] = address.dburi
     server = marimo.create_asgi_app()
     server = server.with_app(path="/", root=cratedb_toolkit.cfr.marimo.__file__)
@@ -238,7 +238,7 @@ cli.add_command(info, name="info")
 @click.pass_context
 def record(ctx: click.Context, once: bool):
     scrub = ctx.meta.get("scrub", False)
-    address = DatabaseAddress.from_string(ctx.meta["cratedb_http_url"] or ctx.meta["cratedb_sqlalchemy_url"])
+    address = DatabaseAddress.from_string(ctx.meta["http_url"] or ctx.meta["sqlalchemy_url"])
     adapter = DatabaseAdapter(dburi=address.dburi, echo=False)
     recorder = InfoRecorder(adapter=adapter, scrub=scrub)
     if once: