+++ cratedb_toolkit/cmd/tail/cli.py
@@ -5,18 +5,15 @@ import click
 
 from cratedb_toolkit.cmd.tail.main import TableTailer
 from cratedb_toolkit.model import TableAddress
+from cratedb_toolkit.option import option_sqlalchemy_url
 from cratedb_toolkit.util.cli import boot_click
 from cratedb_toolkit.util.database import DatabaseAdapter
 
 logger = logging.getLogger(__name__)
 
-cratedb_sqlalchemy_option = click.option(
-    "--cratedb-sqlalchemy-url", envvar="CRATEDB_SQLALCHEMY_URL", type=str, required=False, help="CrateDB SQLAlchemy URL"
-)
-
 
 @click.command()
-@cratedb_sqlalchemy_option
+@option_sqlalchemy_url
 @click.option(
     "--lines", "-n", type=int, required=False, default=10, help="Displays n last lines of the input. Default: 10"
 )
@@ -32,7 +29,7 @@ cratedb_sqlalchemy_option = click.option(
 @click.pass_context
 def cli(
     ctx: click.Context,
-    cratedb_sqlalchemy_url: str,
+    sqlalchemy_url: str,
     resource: str,
     lines: int,
     format_: str,
@@ -44,11 +41,11 @@ def cli(
     """
     A polling tail implementation for database tables.
     """
-    if not cratedb_sqlalchemy_url:
+    if not sqlalchemy_url:
         logger.error("Unable to operate without database address")
         sys.exit(1)
     boot_click(ctx, verbose, debug)
-    adapter = DatabaseAdapter(dburi=cratedb_sqlalchemy_url)
+    adapter = DatabaseAdapter(dburi=sqlalchemy_url)
     # TODO: Tail multiple tables.
     if len(resource) > 1:
         raise NotImplementedError("`ctk tail` currently implements tailing a single table only")